+++ servicex_app/servicex_app/models.py
@@ -168,6 +168,7 @@ class TransformRequest(db.Model):
     request_id = db.Column(db.String(48), unique=True, nullable=False, index=True)
     title = db.Column(db.String(128), nullable=True)
     submit_time = db.Column(db.DateTime, nullable=False)
+    archived = db.Column(db.Boolean, nullable=False, default=False)
     finish_time = db.Column(db.DateTime, nullable=True)
     did = db.Column(db.String(512), unique=False, nullable=False)
     did_id = db.Column(db.Integer, unique=False, nullable=False)
@@ -321,6 +322,10 @@ class TransformRequest(db.Model):
     def results(self) -> List['TransformationResult']:
         return TransformationResult.query.filter_by(request_id=self.request_id).all()
 
+    def truncate_results(self):
+        TransformationResult.query.filter_by(request_id=self.request_id).delete()
+        db.session.commit()
+
     @property
     def all_files(self) -> List['DatasetFile']:
         return DatasetFile.query.filter_by(dataset_id=self.did_id).all()
@@ -359,7 +364,6 @@ class TransformationResult(db.Model):
     __tablename__ = 'transform_result'
 
     id = db.Column(db.Integer, primary_key=True)
-    did = db.Column(db.String(512), unique=False, nullable=False)
     file_id = db.Column(db.Integer, ForeignKey('files.id'))
     file_path = db.Column(db.String(512), unique=False, nullable=False)
     request_id = db.Column(db.String(48), unique=False, nullable=False)
@@ -378,7 +382,6 @@ class TransformationResult(db.Model):
         return {
             'id': x.id,
             'request-id': x.request_id,
-            'did': x.did,
             'file-id': x.id,
             'file-path': x.file_path,
             'transform_status': x.transform_status,
@@ -393,7 +396,7 @@ class TransformationResult(db.Model):
         db.session.flush()
 
 
-class DatasetStatus(Enum):
+class DatasetStatus(str, Enum):
     created = "created"
     looking = "looking"
     complete = "complete"
@@ -403,7 +406,7 @@ class Dataset(db.Model):
     __tablename__ = 'datasets'
 
     id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(1024), unique=True, nullable=False, index=True)
+    name = db.Column(db.String(1024), unique=False, nullable=False, index=True)
     last_used = db.Column(db.DateTime, nullable=False)
     last_updated = db.Column(db.DateTime, nullable=True)
     did_finder = db.Column(db.String(64), nullable=False)
@@ -411,6 +414,7 @@ class Dataset(db.Model):
     size = db.Column(db.BigInteger, default=0, nullable=True)
     events = db.Column(db.BigInteger, default=0, nullable=True)
     lookup_status = db.Column(db.Enum(DatasetStatus), nullable=False)
+    stale = db.Column(db.Boolean, default=False, nullable=False)
     files = relationship("DatasetFile", back_populates="dataset")
 
     def save_to_db(self):
@@ -428,17 +432,32 @@ class Dataset(db.Model):
             'events': self.events,
             'last_used': str(self.last_used.strftime(iso_fmt)),
             'last_updated': str(self.last_updated.strftime(iso_fmt)),
-            'lookup_status': self.lookup_status
+            'lookup_status': self.lookup_status,
+            'is_stale': self.stale
         }
         return result_obj
 
     @classmethod
     def find_by_name(cls, name) -> Optional['Dataset']:
-        return cls.query.filter_by(name=name).first()
+        return cls.query.filter_by(name=name, stale=False).first()
 
     @classmethod
     def find_by_id(cls, id) -> Optional['Dataset']:
         return cls.query.get(id)
+
+    @classmethod
+    def get_by_did_finder(cls, did_finder, show_deleted: bool = False) -> List[Dataset]:
+        if show_deleted:
+            return cls.query.filter_by(did_finder=did_finder)
+        else:
+            return cls.query.filter_by(did_finder=did_finder, stale=False)
+
+    @classmethod
+    def get_all(cls, show_deleted: bool = False) -> List[Dataset]:
+        if show_deleted:
+            return cls.query.all()
+        else:
+            return cls.query.filter_by(stale=False)
 
 
 class DatasetFile(db.Model):
@@ -454,6 +473,15 @@ class DatasetFile(db.Model):
     file_events = db.Column(db.BigInteger, nullable=True)
     paths = db.Column(db.Text(), unique=False, nullable=False)
     dataset = relationship("Dataset", back_populates="files")
+
+    def to_json(self):
+        return {
+            'id': self.id,
+            'adler32': self.adler32,
+            'file_size': self.file_size,
+            'file_events': self.file_events,
+            'paths': self.paths
+        }
 
     def save_to_db(self):
         db.session.add(self)