+++ servicex_app/servicex_app/resources/datasets/delete_dataset.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019, IRIS-HEP
+# Copyright (c) 2024, IRIS-HEP
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -25,33 +25,30 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+from servicex_app.decorators import auth_required
+from servicex_app.models import Dataset
+from servicex_app.resources.servicex_resource import ServiceXResource
 
 
-class ObjectStoreManager:
+class DeleteDataset(ServiceXResource):
+    @auth_required
+    def delete(self, dataset_id):
+        dataset = Dataset.find_by_id(dataset_id)
 
-    def __init__(self, url, username, password, use_https=False):
-        from minio import Minio
-        self.minio_client = Minio(endpoint=url, access_key=username,
-                                  secret_key=password, secure=use_https)
+        if not dataset:
+            return {
+                'message': f'Dataset {dataset_id} not found'
+            }, 404
 
-    def create_bucket(self, bucket_name):
-        self.minio_client.make_bucket(bucket_name)
+        if dataset.stale:
+            return {
+                'message': f'Dataset {dataset_id} has already been deleted'
+            }, 400
 
-    def list_buckets(self):
-        return self.minio_client.list_buckets()
+        dataset.stale = True
+        dataset.save_to_db()
 
-    def delete_bucket_and_contents(self,  bucket_name):
-        if not self.minio_client.bucket_exists(bucket_name):
-            print(f"Bucket '{bucket_name}' does not exist. Nothing to delete.")
-            return
-
-        # List all objects in the bucket
-        objects = self.minio_client.list_objects(bucket_name, recursive=True)
-
-        # Remove each object
-        for obj in objects:
-            self.minio_client.remove_object(bucket_name, obj.object_name)
-
-        # Remove the bucket itself
-        self.minio_client.remove_bucket(bucket_name)
-        print(f"Bucket '{bucket_name}' deleted successfully.")
+        return {
+            'dataset-id': dataset_id,
+            'stale': True
+        }