+++ servicex_app/servicex_app/resources/transformation/submit.py
@@ -176,6 +176,7 @@ class SubmitTransformationRequest(ServiceXResource):
             request_rec = TransformRequest(
                 request_id=str(request_id),
                 title=args.get("title"),
+                archived=False,
                 did=dataset_manager.name,
                 did_id=dataset_manager.id,
                 submit_time=datetime.now(tz=timezone.utc),